+++ src/calender/ChihangCalender/util/index.ts
@@ -1,0 +1,2 @@
+export const makeCustomArray = (length: number) =>
+  Array.from({ length }, (noValue: undefined, index: number) => index);
