+++ src/routes/Routing.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import DetailPage from 'pages/detailPage/DetailPage';
 import MainPage from 'pages/mainPage/MainPage';
 import StatusPage from 'pages/statusPage/StatusPage';
-import Test from 'calender/ChihangCalender/test';
+import CalendarLayout from 'calender/ChihangCalender/CalendarLayout';
 import Path from './Path';
 
 function Routing() {
@@ -12,7 +12,7 @@ function Routing() {
       <Routes>
         <Route path="/" element={<MainPage />} />
         <Route path={Path.status} element={<StatusPage />} />
-        <Route path={Path.detail} element={<Test />} />
+        <Route path={Path.detail} element={<CalendarLayout />} />
       </Routes>
     </BrowserRouter>
   );
